Split SaddleConnections and SaddleConnectionsIterator
