+++ libflatsurf/test/saddle_connections.test.cc
@@ -30,6 +30,7 @@
 #include "../flatsurf/half_edge.hpp"
 #include "../flatsurf/saddle_connection.hpp"
 #include "../flatsurf/saddle_connections.hpp"
+#include "../flatsurf/saddle_connections_iterator.hpp"
 #include "../flatsurf/vector.hpp"
 
 #include "surfaces.hpp"
